Tweak calculation of FS_ACCURENCY
Before this accuracy detection code would end up leaving `FS_ACCURENCY`
equal to 10000 or 2000 on my machine (depending on whether the
timestamp was something like 1455197568000 or 1455197569000). Which was
pretty far away from the actual precision of one second… and concerning
to me if that would make us cache too few things.

So my attempt here was to assume that FS accuracy was always an order
of magnitude—1, 10, 100, or 1000 milliseconds—and be a little more
aggressive to set the `FS_ACCURENCY` lower. For example, in the new
code, any time a timestamp is seen with 3 trailing zeros (`mtime %
10000`), I assume that the precision is 1000ms.